completed all the failsafe functionality for user input

+++ lib/models/User.rb
@@ -19,10 +19,17 @@ class User < ActiveRecord::Base
     def add_to_my_favorites
         Hike.browse_hikes
         puts "Which hike would you like to add?"
-        user_choice = gets.strip.to_i
-        selected_hike = Hike.all[user_choice - 1]
-        new_favorite = FavoritesList.create(hike: selected_hike, user: $user)
-        show_my_favorites
+        loop do
+            user_choice = gets.strip.to_i
+            if user_choice > 0 && user_choice <= Hike.all.length  
+                selected_hike = Hike.all[user_choice - 1]
+                new_favorite = FavoritesList.create(hike: selected_hike, user: $user)
+                show_my_favorites
+                break
+            else
+                puts "Please choose again - invalid entry"
+            end
+        end
     end
     
     def my_favorite_hikes
@@ -46,12 +53,18 @@ class User < ActiveRecord::Base
         puts "\n\n"
         if my_hikes.length != 0
             puts "Would you like to edit your favorites? Y/N"
-            user_input = gets.strip.upcase
-            case user_input
-            when "Y" 
-                edit_my_favorites
-            when "N"
-                puts "Okay! No problem."
+            loop do
+                user_input = gets.strip.upcase
+                case user_input
+                when "Y" 
+                    edit_my_favorites
+                    break
+                when "N"
+                    puts "Okay! No problem."
+                    break
+                else
+                    puts "Invalid entry:  Please enter Y/N" 
+                end
             end
         end
         
@@ -60,19 +73,20 @@ class User < ActiveRecord::Base
     def edit_my_favorites
         my_hikes = my_favorite_hikes
         puts "Which number would you like to remove?"
-        hike_selection = gets.strip.to_i
-        hike_to_delete_from_my_favorites = my_hikes[hike_selection - 1]
-        
-        favorites_list_entry_to_delete = FavoritesList.find_by(user_id: $user.id, hike_id: hike_to_delete_from_my_favorites.id)
-        
-        favorites_list_entry_to_delete.destroy
-        
+        loop do 
+            hike_selection = gets.strip.to_i
+            if hike_selection > 0 && hike_selection <= my_hikes.length
+                hike_to_delete_from_my_favorites = my_hikes[hike_selection - 1]
+                favorites_list_entry_to_delete = FavoritesList.find_by(user_id: $user.id, hike_id: hike_to_delete_from_my_favorites.id)
+                favorites_list_entry_to_delete.destroy
+                break
+            else    
+                puts "Please choose again - invalid entry"
+            end
+        end
         puts "Here is your updated favorites list:\n\n"
         show_my_favorites
-        
     end
-
-
-
+          
 end
